+++ src/Filter/Validation/Validator/StringLength/Min.php
@@ -27,6 +27,11 @@ use function strlen;
  * The test is passed if for a string's length L, min<=L, i.e. L must
  * be at least min.
  *
+ * The "included" option is true by default. Set the option to false
+ * for min<L, i.e. L must be more than min. The "includedMinimum" option
+ * is an alias of "included". If you set the two options, "included" has
+ * precedence.
+ *
  * ```php
  * use Phalcon\Filter\Validation;
  * use Phalcon\Filter\Validation\Validator\StringLength\Min;
@@ -39,7 +44,7 @@ use function strlen;
  *         [
  *             "min"     => 2,
  *             "message" => "We want more than just their initials",
- *             "included" => true
+ *             "included" => false
  *         ]
  *     )
  * );
@@ -115,7 +120,18 @@ class Min extends AbstractValidator
             $minimum = $minimum[$field];
         }
 
-        $included = $this->getOption("included");
+        /**
+         * "includedMinimum" is an alias of "included". The minimum is
+         * inclusive if no option is set. hasOption() uses isset(), thus a
+         * null value also counts as not set
+         */
+        $included = true;
+
+        if ($this->hasOption("included")) {
+            $included = $this->getOption("included");
+        } elseif ($this->hasOption("includedMinimum")) {
+            $included = $this->getOption("includedMinimum");
+        }
 
         if (is_array($included)) {
             $included = (bool)$included[$field];